top up settings

+++ app/Models/TopUpVendor.php
@@ -6,5 +6,5 @@ class TopUpVendor extends Model
 {
     protected $guarded = ['id'];
     protected $table = 'topup_vendors';
-    protected $casts = ['sheba_commission' => 'double', 'agent_commission' => 'double', 'is_published' => 'double'];
+    protected $casts = ['sheba_commission' => 'double', 'agent_commission' => 'double', 'is_published' => 'int'];
 }
